Harden runtime logic and add backwards compat errors
