+++ UVTD/xmake.lua
@@ -1,6 +1,6 @@
 local projectName = "UVTD"
 
-add_requires("raw_pdb", { debug = is_mode_debug() })
+add_requires("raw_pdb", { debug = is_mode_debug(), configs = {runtimes = get_mode_runtimes()} })
 
 target(projectName)
     set_kind("binary")
