+++ deps/first/Profiler/xmake.lua
@@ -2,8 +2,8 @@ local projectName = "Profiler"
 
 includes("deps/Tracy")
 
-add_requires("Tracy", { optional = true, debug = is_mode_debug() })
-add_requires("Superluminal", { system = true, optional = true, debug = is_mode_debug() })
+add_requires("Tracy", { optional = true, debug = is_mode_debug(), configs = {runtimes = get_mode_runtimes()}})
+add_requires("Superluminal", { system = true, optional = true, debug = is_mode_debug(), configs = {runtimes = get_mode_runtimes()} })
 
 option("profilerFlavor")
     set_default("Tracy")
