+++ deps/third/xmake.lua
@@ -6,6 +6,6 @@ includes("imgui")
 includes("glad")
 includes("raw_pdb")
 
-add_requires("zycore v1.5.0", { debug = is_mode_debug() })
-add_requires("zydis v4.1.0", { debug = is_mode_debug() })
-add_requires("polyhook_2", { debug = is_mode_debug() })
+add_requires("zycore v1.5.0", { debug = is_mode_debug(), configs = {runtimes = get_mode_runtimes()} })
+add_requires("zydis v4.1.0", { debug = is_mode_debug(), configs = {runtimes = get_mode_runtimes()} })
+add_requires("polyhook_2", { debug = is_mode_debug(), configs = {runtimes = get_mode_runtimes()} })
